Added a Saving system using Binary files
Player can now save the game, load the game (not completely finished, need to include inventory) and can quit the game via pause menu.

+++ 2D-RPG/Assets/Scripts/Components/ExperienceComponent.cs
@@ -10,7 +10,7 @@ public class ExperienceComponent : MonoBehaviour
 	public int baseExp = 100;
 	public int expToLevel = 100;
 	public float expIncrease = 1.7f;
-    public float totalExp;
+    public int totalExp;
 
 	public Slider expBar;
     public Text currentExpDisplay;
